+++ gems/aws-sdk-cloudcontrolapi/lib/aws-sdk-cloudcontrolapi.rb
@@ -55,7 +55,7 @@ module Aws::CloudControlApi
   autoload :EndpointProvider, 'aws-sdk-cloudcontrolapi/endpoint_provider'
   autoload :Endpoints, 'aws-sdk-cloudcontrolapi/endpoints'
 
-  GEM_VERSION = '1.36.0'
+  GEM_VERSION = '1.37.0'
 
 end