+++ gems/aws-sdk-eks/lib/aws-sdk-eks.rb
@@ -55,7 +55,7 @@ module Aws::EKS
   autoload :EndpointProvider, 'aws-sdk-eks/endpoint_provider'
   autoload :Endpoints, 'aws-sdk-eks/endpoints'
 
-  GEM_VERSION = '1.129.0'
+  GEM_VERSION = '1.130.0'
 
 end