+++ gems/aws-sdk-mediaconnect/lib/aws-sdk-mediaconnect.rb
@@ -55,7 +55,7 @@ module Aws::MediaConnect
   autoload :EndpointProvider, 'aws-sdk-mediaconnect/endpoint_provider'
   autoload :Endpoints, 'aws-sdk-mediaconnect/endpoints'
 
-  GEM_VERSION = '1.76.0'
+  GEM_VERSION = '1.77.0'
 
 end