+++ gems/aws-sdk-qbusiness/lib/aws-sdk-qbusiness/async_client.rb
@@ -630,7 +630,7 @@ module Aws::QBusiness
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-qbusiness'
-      context[:gem_version] = '1.29.0'
+      context[:gem_version] = '1.30.0'
       Seahorse::Client::Request.new(handlers, context)
     end