+++ gems/aws-sdk-qldbsession/lib/aws-sdk-qldbsession.rb
@@ -54,7 +54,7 @@ module Aws::QLDBSession
   autoload :EndpointProvider, 'aws-sdk-qldbsession/endpoint_provider'
   autoload :Endpoints, 'aws-sdk-qldbsession/endpoints'
 
-  GEM_VERSION = '1.50.0'
+  GEM_VERSION = '1.51.0'
 
 end