+++ gems/aws-sdk-route53/lib/aws-sdk-route53.rb
@@ -55,7 +55,7 @@ module Aws::Route53
   autoload :EndpointProvider, 'aws-sdk-route53/endpoint_provider'
   autoload :Endpoints, 'aws-sdk-route53/endpoints'
 
-  GEM_VERSION = '1.109.0'
+  GEM_VERSION = '1.110.0'
 
 end