+++ gems/aws-sdk-sesv2/lib/aws-sdk-sesv2.rb
@@ -54,7 +54,7 @@ module Aws::SESV2
   autoload :EndpointProvider, 'aws-sdk-sesv2/endpoint_provider'
   autoload :Endpoints, 'aws-sdk-sesv2/endpoints'
 
-  GEM_VERSION = '1.70.0'
+  GEM_VERSION = '1.71.0'
 
 end